+++ sources/www/js/objects/acquisition-manager.js
@@ -31,8 +31,11 @@ AcquisitionManager.factory = function (json) {
  */
 AcquisitionManager.getLastAcquisition = function (keywords, additionalInfos) {
    var d = new $.Deferred();
-   var getAdditionInfo = ["lastValue", "lastValueDate"];
+   var getAdditionInfo = ["lastValue", "lastValueDate","exist"];
 
-   getAdditionInfo = getAdditionInfo.concat(additionalInfos);
+   //append additional informations
+   if(additionalInfos) {
+	  getAdditionInfo = getAdditionInfo.concat(additionalInfos);
+   }
    return KeywordManager.getInformation(keywords, getAdditionInfo);
 }