+++ sources/www/js/widget-engine.js
@@ -425,7 +425,7 @@ function updateWidgetsPolling(pageId) {
              //we signal the new acquisition to the widget if the widget support the method
              $.each(pageId.widgets, function (widgetIndex, widget) {
                 if ($.inArray(acquisition.keywordId, widget.getlastValue)!=-1){
-                   if (isNullOrUndefined(acquisition.error)){
+                   if (parseBool(acquisition.exist)){
                       if (widget.viewModel.onNewAcquisition !== undefined)
                          widget.viewModel.onNewAcquisition(acquisition.keywordId, acquisition);
                    }else{ // we desactivate the widget
@@ -469,8 +469,8 @@ function updateWidgetPolling(widget) {
             })
             .fail(function (error) {
                 notifyError($.t("objects.generic.errorGetting",{
-                        objectName: "last acquisition for widget = " + widget.id
-                    }),error);
+                   objectName: "last acquisition for widget = " + widget.id
+                   }),error);
              d.reject(error);
           });
        } else 
@@ -486,11 +486,10 @@ function updateWidgetPollingByKeywordsId(keywords, additionnalInfo) {
     
     if (!isNullOrUndefined(keywords)) {
        if (keywords!=0){
-          console.log (additionnalInfo);
           AcquisitionManager.getLastAcquisition(keywords, additionnalInfo)
           .done(d.resolve)
           .fail(function (error) {
-             notifyError($.t("objects.generic.errorGetting", { objectName: "last acquisition for widget = " + infos }), error);
+             notifyError($.t("objects.widgetManager.errorDuringGettingKeywordsInformation"), error);
              d.reject(error);
           });
        } else