+++ sources/www/widgets/chart/changelog.md
@@ -1,8 +1,13 @@
+### 1.2.0-beta.2
+* Fix #508 - Error when displaying some charts
+
 ### 1.2.0-beta.1
 * NewFeature #466 - Adapt units if needed
 * NewFeature #485 - Adapt unit for bit/s display
 * Fix #389 - Bad display after computer wake up
 * Fix #500 - error log every minutes for only few tabs
+* Fix #503 - Export file - hour format in GMT hour
+* Fix #505 - Mismatch export function when more than one chart in a page
 
 ### 1.1.0
 * fix #462 : doesn't display the first value for some summary values